+++ tests/networked/test_latest_tag.expected.nix
@@ -1,0 +1,6 @@
+fetchFromGitHub {
+  owner = "expipiplus1";
+  repo = "update-nix-fetchgit";
+  rev = "e38aedf11616eb74d06cf99a5244290eef5d404e"; # tags/0.1*
+  sha256 = "1k8mxdp789lp3pwcm8i8dz65yjcsx5ijz1fmiwghwr4kzsa6z20c";
+}